pyproject.toml
[tool.poetry]
name = "python-grpc"
version = "0.1.0"
description = "gRPC"
authors = ["Rajat Arora <arora_rajat@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio = "^1.71.0"
grpcio-tools = "^1.71.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
